Extract PropertyParser
